.claude/skills/migration-report-dashboard/requirements.txt
# Migration Report Dashboard - Python Dependencies

# Core Dashboard
streamlit==1.31.0
streamlit-aggrid==0.3.4
streamlit-option-menu==0.3.6

# Data Processing
pandas==2.1.4
numpy==1.24.3

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# API & HTTP
requests==2.31.0
httpx==0.26.0
aiohttp==3.9.1

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pymongo==4.6.1

# Code Quality Integration
sonarqube-api==2.2.0
python-sonarqube-api==1.3.4

# Coverage Parsers
coverage==7.4.0
pytest-cov==4.1.0

# Report Generation
reportlab==4.0.9
jinja2==3.1.3
markdown==3.5.2
pdfkit==1.0.0
weasyprint==60.2

# Excel Export
openpyxl==3.1.2
xlsxwriter==3.1.9

# YAML Config
pyyaml==6.0.1
python-dotenv==1.0.0

# Date/Time
python-dateutil==2.8.2
pytz==2023.3

# Utilities
tqdm==4.66.1
colorlog==6.8.0
tabulate==0.9.0

# Schema Comparison
deepdiff==6.7.1
jsondiff==2.0.0

# Performance Testing
locust==2.20.1

# Accessibility Testing
axe-selenium-python==2.1.6
selenium==4.16.0

# API Testing
tavern==2.9.2
